backend/requirements.txt
groq
gtts
langdetect
google.generativeai
praw